+++ Src/Style/FrmStyle.h
@@ -99,9 +99,8 @@ private Q_SLOTS:
     void on_pbCancel_clicked();
     void on_pbBrowse_clicked();
     void on_pbDefault_clicked();
-    
     void on_gpIconTheme_clicked();
-    
+
 private:
     Ui::CFrmStyle *ui;
 };