+++ ChangeLog.md
@@ -7,6 +7,7 @@
   - QTextBrowser replace QTextEdit
   - Package MacOS to end
   - Fix Unable to change "Icon theme" after disabling it. See: https://github.com/KangLin/RabbitRemoteControl/issues/69
+  - Fix the theme to other color when the system theme is dark. See: https://github.com/KangLin/RabbitRemoteControl/pull/68
 
 ### Version v2.3.3
   - CAdminAuthorization: Add pkexec in linux